address review comments (#1055)
Move filter function into existing files
