+++ lib/experimental/new_embed.dart
@@ -5,7 +5,6 @@
 import 'dart:async';
 import 'dart:html' hide Document;
 
-import 'package:dart_pad/experimental/filter_output.dart';
 import 'package:split/split.dart';
 
 import '../completion.dart';
@@ -697,4 +696,16 @@ class NewEmbedContext {
     // TODO(DomesticMouse): implement with CodeMirror integration
     return false;
   }
+}
+
+final RegExp _flutterUrlExp =
+    RegExp(r'(https:[a-zA-Z0-9_=%&\/\-\?\.]+flutter_web\.js)(:\d+:\d+)');
+final RegExp _dartUrlExp =
+    RegExp(r'(https:[a-zA-Z0-9_=%&\/\-\?\.]+dart_sdk\.js)(:\d+:\d+)');
+
+String filterCloudUrls(String trace) {
+  return trace
+      .replaceAllMapped(
+          _flutterUrlExp, (m) => '[Flutter SDK Source]${m.group(2)}')
+      .replaceAllMapped(_dartUrlExp, (m) => '[Dart SDK Source]${m.group(2)}');
 }
