+++ test/experimental/new_embed_test.dart
@@ -64,4 +64,17 @@ void main() {
       expect(editorTab.classes, isNot(contains('selected')));
     });
   });
+  group('filterCloudUrls', () {
+    test('cleans dart SDK urls', () {
+      var trace =
+          '(https://storage.googleapis.com/compilation_artifacts/2.2.0/dart_sdk.js:4537:11)';
+      expect(new_embed.filterCloudUrls(trace), '([Dart SDK Source]:4537:11)');
+    });
+    test('cleans flutter SDK urls', () {
+      var trace =
+          '(https://storage.googleapis.com/compilation_artifacts/2.2.0/flutter_web.js:96550:21)';
+      expect(
+          new_embed.filterCloudUrls(trace), '([Flutter SDK Source]:96550:21)');
+    });
+  });
 }
